incorrect type for market global value fields

+++ models/extern/CMC/GlobalData.go
@@ -8,8 +8,8 @@ import (
 
 // GlobalData is the storage for the JSON from the CMC api
 type GlobalData struct {
-	TotalMarketCapUsd            int64   `json:"total_market_cap_usd"`
-	Total24HVolumeUsd            int64   `json:"total_24h_volume_usd"`
+	TotalMarketCapUsd            float64 `json:"total_market_cap_usd"`
+	Total24HVolumeUsd            float64 `json:"total_24h_volume_usd"`
 	BitcoinPercentageOfMarketCap float64 `json:"bitcoin_percentage_of_market_cap"`
 	ActiveCurrencies             int     `json:"active_currencies"`
 	ActiveAssets                 int     `json:"active_assets"`
